get the XML files from the device
$. getTrees.sh 15.12.13.18
From the input ip address, the script reads the DiscoveryTree.xml, which has the list of XMLs that store the configurations of the printer.
Then it retrieves them, creates the folder called Workplace in the same directory of the script and store the files. 
The purpose is to help the simulation with retrieved XML files in the own repo, need not to connect the device all the time.