Add files via upload
. parse.sh binary_file

parse.sh will take a binary file as an input and produce the content in XML format. It is useful when you think to post large XML file in the binary format over the net, to speed up and in compact secure format. At the source, parse.sh will convert the binary file into XML format. 

If God gives you more time, help this file to improve in handling complicated XML files.