Find required word from the product
Helper script. 
In the model of the localization, the strings displayed at the web page is actually coded with ENUM, in order to achieve the localization. Each ENUM can be mapped to different strings which are written in 28 languages. In the picture, javascript page has only the ENUMs, of which description in 28 languages are in separate file which is written in C header. Mapping is done by framework. Each product model has own string file. We typically support 7-10 products per year, from Enterprise to consumer. 

Then we got an issue, when marketing wanted to change a word to another word in a range of product instantly, 'Ink' to 'Cartridge', for all Enterprise products. They want to know how many words the developer need to re-localize and calculate the schedule and the budget.
Since developer can only know the ENUMs used in the dedicated product, we faced the hard time, to , first, collect the total strings including the word 'Ink'. There are basically 5000+ string