Switch the branch without checkout any files
. justSwitchBranch.sh another_branch_name
Sometimes, when you try to branch out from one branch to another in same repo, Git shows the list of certain files to checkout in prior. You should know the reason for the need to do that!
But sometimes, the list seems too long to manually checked out one by one, and (assuming that) you are lazy to do that or too busy to do that, you can run this script. Important : You are also sure that there is NO file you want to stash or keep it in current branch. You just want all those files in the list to be checked-out themselves and switch to the required branch right away.Note : You can place this script at anywhere in a repo, need not to be placed at the root