Add files via upload
The work requires to sync iorefs [ioref : enum] between the file in server side .h  and the file in client side .js, in which server-side has only sub-set of client-side file.  JS developer forgets to update server side header file. 
This script will help to generate the missing iorefs with relatives enums, ready to update in server side. 
Prior input to this script are two files which one is from .h and one is from .js.
. getio.sh
What is ioref?  long to explain. talk to me. :)