This script helps to find all strings being used in Apps, so that we can save for localization.
$. findUsedStrings.sh
Today master string file has 7910 strings. This script shows that total number of strings being used in apps are actually only 4256. 
When three new languages are requested to localize, now, we have a choice to localize only used strings. Saving lots of money for company. 
This script needs to work with Excel on this result file.