Yearly Calendar on ReactJS
This Yearly Calendar is based on React and can do

+ Highlight the current date.

+ Add the ability to mark a date as one of these categories, by clicking on the date:
• Holiday
• Birthday
• Busy
• Anniversary

Each category should have a different colour. When a date is marked with a category, that date should be
highlighted with that colour (similar to highlighted dates in the above illustration).
Clicking on a date that has a category assigned to it should allow changing or removing the category.

+ Enhance the calendar to support more than the current year. Provide navigation so the user can switch between
different years and a way to navigate to the current year.